Refaktorointia, rekisteröinnin tila.
